department distribution , views

+++ flutter_utils/lib/utils/extensionHelper.dart
@@ -459,7 +459,8 @@ mixin HappyExtensionHelper implements HappyExtensionHelperCallback2{
   }
 
   void sysDeleteHE_ListView(dynamic he_listViewBody,String primaryKey,{Function? successCallback,String dataJson="",
-    String content="Are you sure want to delete ?",DevelopmentMode developmentMode=DevelopmentMode.traditional,TraditionalParam? traditionalParam}){
+    String content="Are you sure want to delete ?",DevelopmentMode developmentMode=DevelopmentMode.traditional,
+    TraditionalParam? traditionalParam,RxBool? loader}){
     CustomAlertUtil(
         callback: (){
           postUIJson(getPageIdentifier(),
@@ -472,11 +473,15 @@ mixin HappyExtensionHelper implements HappyExtensionHelperCallback2{
                 if(successCallback!=null){
                   successCallback(e);
                 }
-                he_listViewBody.updateArrById(primaryKey, e["Table"][0],action: ActionType.deleteById);
+                if(he_listViewBody!=null && e["Table"]!=null && e["Table"].length>0){
+                  he_listViewBody.updateArrById(primaryKey, e["Table"][0],action: ActionType.deleteById);
+                }
+
                 //updateArrById(primaryKey, e["Table"][0], arr,action: ActionType.deleteById,primaryArr:primaryArr );
               },
               developmentMode: developmentMode,
-              traditionalParam:  traditionalParam
+              traditionalParam:  traditionalParam,
+            loader: loader
           );
         },
         cancelCallback: (){
